refactor: move pull request fetching into separate fn
