+++ merge-by/scripts/build-table-and-notify.js
@@ -1,4 +1,5 @@
 import { QUERIES, MUTATIONS } from "./graphql";
+import { getPullRequests } from "./promises";
 
 /**
  * Builds a row for the Markdown table given the GitHub repo owner, repo name and pull request.
@@ -15,11 +16,14 @@ import { QUERIES, MUTATIONS } from "./graphql";
  * @returns
  */
 const buildTableRow = (owner, repo, pr) =>
-  `| [#${pr.number}](https://github.com/${owner}/${repo}/pull/${pr.number
-  }) | [**${pr.title.trim()}**](https://github.com/${owner}/${repo}/pull/${pr.number
-  }) | ${pr.author} | ${pr.mergeBy ?? "N/A"} | ${pr.hasReviews && pr.mergeable !== false
-    ? ":white_check_mark:"
-    : ":white_large_square:"
+  `| [#${pr.number}](https://github.com/${owner}/${repo}/pull/${
+    pr.number
+  }) | [**${pr.title.trim()}**](https://github.com/${owner}/${repo}/pull/${
+    pr.number
+  }) | ${pr.author} | ${pr.mergeBy ?? "N/A"} | ${
+    pr.hasReviews && pr.mergeable !== false
+      ? ":white_check_mark:"
+      : ":white_large_square:"
   } |`;
 
 const TABLE_HEADER = `
@@ -163,69 +167,7 @@ const notifyUsers = async ({
  */
 const fetchPullRequests = async ({ dayJs, github, owner, repo, today }) => {
   const nextWeek = today.add(7, "day");
-  return (
-    await Promise.all(
-      (
-        await github.rest.pulls.list({
-          owner,
-          repo,
-          state: "open",
-        })
-      )?.data
-        .filter((pr) => !pr.draft)
-        .map(async (pr) => {
-          const comments = (
-            await github.rest.issues.listComments({
-              owner,
-              repo,
-              issue_number: pr.number,
-            })
-          ).data;
-          // Attempt to parse the merge-by date from the bot comment
-          const existingComment = comments?.find(
-            (comment) =>
-              comment.user.login === "github-actions[bot]" &&
-              comment.body.includes("**📅 Suggested merge-by date:")
-          );
-
-          const reviews = (
-            await github.rest.pulls.listReviews({
-              owner,
-              repo,
-              pull_number: pr.number,
-            })
-          ).data;
-
-          const hasTwoReviews =
-            reviews.reduce(
-              (all, review) => (review.state === "APPROVED" ? all + 1 : all),
-              0
-            ) >= 2;
-
-          // Filter out reviewers if they have already reviewed and approved the pull request
-          const reviewersNotApproved = pr.requested_reviewers.filter(
-            (reviewer) =>
-              reviews.find(
-                (review) =>
-                  review.state === "APPROVED" &&
-                  reviewer.login === review.user.login
-              ) == null
-          );
-
-          return {
-            number: pr.number,
-            title: pr.title,
-            author: pr.user.login,
-            hasReviews: hasTwoReviews,
-            mergeable: pr.mergeable,
-            reviewers: reviewersNotApproved,
-            mergeBy: existingComment?.body
-              .substring(existingComment.body.lastIndexOf("*") + 1)
-              .trim(),
-          };
-        })
-    )
-  )
+  return (await Promise.all(getPullRequests({ github })))
     .filter((pr) => {
       if (pr.mergeBy == null) {
         return true;
